server update

+++ src/server/game/Loot/LootItemStorage.h
@@ -63,9 +63,9 @@ public:
     void RemoveMoney();
     void RemoveItem(uint32 itemId, uint32 count, uint32 itemIndex);
 
-    uint32 GetContainer() const { return _containerId; }
-    uint32 GetMoney() const { return _money; }
-    StoredLootItemContainer const& GetLootItems() const { return _lootItems; }
+        uint64 GetContainer() const { return _containerId; }
+        uint32 GetMoney() const { return _money; }
+        StoredLootItemContainer const& GetLootItems() const { return _lootItems; }
 
 private:
     StoredLootItemContainer _lootItems;
